Lagt inn generell felhåndtering i plugin-koden. Endre navn på plugin og metoder. Redusert minnelekkasje.
